+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Run the examples in the continuous integration pipeline #26
 * Automatically create a Github release when a version tag is pushed #28
 * Add `number_of_qubits` to the `quantum_circuit` job payload #29
+* Fix substitution circuit for wrapping the `Rxx` angles #30
 
 ## qiskit-aqt-provider v0.8.1